refactoring misc kwargs in play; also updating tools/ perf_test.py profile_test.py to use these kwargs
