+++ basic_engine/tools/profile_test.py
@@ -1,10 +1,12 @@
+import sys
 import cProfile
 import pstats
+sys.path.append('../')
 
-from main import Game
+from src.main import Game
 
-ss = "1. h7 f8 2. b1 c1 3. g5 e5 4. b2 c2 5. h6 f4 6. b3 c3 7. h5 h6 8. b4 c4 9. h6 h5 10. b5 c5"
-ss_long = '1. g1 e1 2. b1 d1 3. g2 e2 4. b3 d3 5. e2 d3 6. b6 d6 7. g5 e5 8. a2 c3 9. h4 d8 10. b7 c7 11. h6 c1 12. a1 c1 13. h1 f1 14. a6 c8 15. h7 f6 16. b2 d2 17. h3 g2 18. a5 a6 19. e1 d2 20. c8 d7 21. d8 c7 22. d7 e6 23. h5 h7 24. b8 c8 25. g3 e3 26. e6 b3 27. g7 e7 28. c3 e4 29. c7 b7 30. a6 a5 31. b7 c7 32. c1 c7 33. d2 c2 34. b4 d4 35. f6 e4 36. d6 e5 37. h2 f3 38. c8 d8 39. f1 d1 40. c7 c4 '
+ss = "1. g1 h3 2. a7 a6 3. e2 e4 4. b7 b6 5. f1 d3 6. c7 c6 7. e1 f1 8. d7 d6 9. f1 e1 10. e7 e6"
+
 
 #os.chdir("c:/users/wsutt/desktop/files/exercises/chess/basic_engine/")
 
@@ -17,7 +19,7 @@ def run_profiles():
 
     cmd =  """game = Game(s_instructions = ss); """
     cmd += """game.play(filter_check_naive=False """
-    cmd += """          ,king_in_check_test_copy_apply=True """
+    cmd += """          ,filter_check_test_copy_apply=True """
     cmd += """          )"""
     cProfile.run( cmd, 'output_profile_2')
     
@@ -27,7 +29,7 @@ def run_profiles():
 
     cmd =  """game = Game(s_instructions = ss); """
     cmd += """game.play(filter_check_naive=False """
-    cmd += """          ,king_in_check_optimal=True """
+    cmd += """          ,filter_check_test_copy_opt=True """
     cmd += """          )"""
     cProfile.run( cmd, 'output_profile_4')
